refactor: перестроена архитектура проекта по FSD
