+++ my-app/src/entities/errorMessage.tsx
