+++ my-app/src/entities/type.ts
