+++ my-app/src/features/movieCard.tsx
