+++ my-app/src/features/videoPlayer.tsx
