+++ my-app/src/main.tsx
@@ -1,9 +1,9 @@
 import "./index.css"
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import App from "./app/App";
 import { Provider } from "react-redux";
-import { store } from "./store";
+import { store } from "./app/store";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
