+++ my-app/src/shared/useAppDispatch.tsx
@@ -1,4 +1,4 @@
 import { useDispatch } from "react-redux";
-import type { AppDispatch } from "../store";
+import type { AppDispatch }  from "../app/store"
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
