+++ my-app/src/shared/useAppSelector.tsx
@@ -1,4 +1,4 @@
 import type{ TypedUseSelectorHook, } from "react-redux";
 import { useSelector } from "react-redux"; 
-import type { RootState } from "@/store";
+import type { RootState } from "@/app/store";
 export const useAppSelector:TypedUseSelectorHook<RootState> = useSelector
