+++ my-app/src/widgets/loader.tsx
